Eliminated a compilation warning.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@1937 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/vncviewer/FileTransferTypes.cpp
@@ -40,5 +40,5 @@ Time70ToFiletime(unsigned int time70, FILETIME *pftime)
 {
     LONGLONG ll = Int32x32To64(time70, 10000000) + 116444736000000000;
     pftime->dwLowDateTime = (DWORD) ll;
-    pftime->dwHighDateTime = (ll >> 32);
+    pftime->dwHighDateTime = (DWORD) (ll >> 32);
 }
